chore: add aws-us-east-1 cluster profile
